Coding Update 2
Added a couple more elements to the coding page. Looked into a couple more ways to change the sidebar. I'm currently leaning towards making it a separate html page with an iframe, but we'll see. In the meantime, manually changed the sidebar to say "graphic design" instead of "social media".